\name{arnold-package}
\alias{arnold-package}
\alias{arnold}
\docType{package}
\title{
\packageTitle{arnold}
}
\description{
\packageDescription{arnold}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{arnold}
\packageIndices{arnold}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{arnold}

Maintainer: \packageMaintainer{arnold}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
